requirements.txt
django-channels
channels-redis
 